# backend/API/wwwroot/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Filmovi API",
    "version": "1.0.0"
  },
  "contact": {
    "name": "Sime Braica",
    "email": "sime@sime.com"
  },
  "license": {
    "name": "CC0",
    "url": "https://creativecommons.org/public-domain/cc0/"
  },
  "paths": {
    "/api/Movie": {
      "get": {
        "summary": "Dohvati sve filmove",
        "responses": {
          "200": {
            "description": "Action was succesfull",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Action was succesfull"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MovieDTO"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Action was not succesfull, resource does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Action was not succesfull, resource does not exist"
                    },
                    "data": {
                      "type": "null"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Action was not succesfull, internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Action was not succesfull, internal server error"
                    },
                    "data": {
                      "type": "null"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Movie/MoviesByNaziv": {
      "get": {
        "summary": "Dohvati filmove po nazivu",
        "parameters": [
          {
            "name": "naziv",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Action was succesfull",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Action was succesfull"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MovieDTO"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Action was not succesfull, search term invalid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Action was not succesfull, search term invalid"
                    },
                    "data": {
                      "type": "null"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/Movie/MoviesByZanr": {
      "get": {
        "summary": "Dohvati filmove po žanru",
        "parameters": [
          {
            "name": "zanr",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/Movie/MoviesByWildcard": {
      "get": {
        "summary": "Dohvati filmove po wildcardu",
        "parameters": [
          {
            "name": "wildcard",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/Movie/MoviesById": {
      "get": {
        "summary": "Dohvati film po ID-u",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/Movie/CreateMovie": {
      "post": {
        "summary": "Kreiraj novi film",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostMovieDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Action was succesfull, resource created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Action was succesfull, resource created"
                    },
                    "data": {
                      "type": "null"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/Movie/DeleteMovie": {
      "delete": {
        "summary": "Izbriši film",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/Movie/UpdateMovie": {
      "put": {
        "summary": "Ažuriraj film",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    }
  }
}